dir renamed. Replay schema constructed via init.py
