+++ __main__.py
@@ -1,0 +1,17 @@
+from database.config import SessionLocal
+from database.init import *
+
+import sc2reader
+from sc2reader.engine.plugins import (
+    SelectionTracker,
+    APMTracker,
+    ContextLoader,
+    GameHeartNormalizer,
+)
+
+sc2reader.engine.register_plugin(SelectionTracker())
+sc2reader.engine.register_plugin(APMTracker())
+sc2reader.engine.register_plugin(ContextLoader())
+sc2reader.engine.register_plugin(GameHeartNormalizer())
+
+replay = sc2reader.load_replay("example_5.SC2Replay")
