+++ database/raw/events/BASIC_COMMAND_EVENT.sql
