+++ database/raw/events/CHAT_EVENT.sql
