+++ database/raw/events/PLAYER_LEAVE_EVENT.sql
