+++ database/raw/events/PLAYER_SETUP_EVENT.sql
