+++ database/raw/events/UNIT_DIED_EVENT.sql
