+++ database/raw/events/UPGRADE_COMPLETE_EVENT.sql
